app/backend/requirements.txt
h5py
tensorflow-cpu
passlib
sqlmodel
pydantic==2.3.0
pydantic-settings==2.0.3
asyncpg==0.28.0
psycopg2-binary
psycopg-binary==3.1.10
pandas
asyncio
fastapi
uvicorn
pika